+++ src/edu/saddleback/cs4b/Backend/Messages/BaseMessage.java
@@ -1,7 +1,5 @@
 package edu.saddleback.cs4b.Backend.Messages;
 
-import edu.saddleback.cs4b.Backend.Utilitys.MsgTypes;
-
 import java.io.Serializable;
 import java.time.LocalDateTime;
