+++ src/edu/saddleback/cs4b/Backend/Messages/SignInMessage.java
@@ -1,6 +1,5 @@
 package edu.saddleback.cs4b.Backend.Messages;
 
-import edu.saddleback.cs4b.Backend.Utilitys.MsgTypes;
 import edu.saddleback.cs4b.Backend.Utilitys.User;
 
 public class SignInMessage extends BaseMessage {
